+++ src/main/java/mate/academy/database/user/UserDao.java
@@ -8,7 +8,7 @@ public interface UserDao {
     int addUser(User user);
     int removeUser(User user);
     Optional<User> getUserByLogin(String login);
-    Optional<User> getUserById(int id);
+    Optional<User> getUserById(Long id);
     Optional<User> getUserByMail(String mail);
     int editUser(User user);
     List<User> getUsers();
